Added an abstract class
