+++ CSharpFundamentals/XUnitCSharpFundamentals/TypeTests.cs
@@ -65,21 +65,21 @@ namespace XUnitCSharpFundamentals
         [Fact]
         public void CSharpIsPassedByValue()
         {
-            var book1 = GetBook("Book 1");
+            var book1 = GetBook("InMemoryBook 1");
             GetBookSetName(ref book1, "New Jersey");
 
             Assert.Equal("New Jersey", book1.Name);
         }
 
-        void GetBookSetName(ref Book book, string name)
+        void GetBookSetName(ref InMemoryBook book, string name)
         {
-            book = new Book(name);
+            book = new InMemoryBook(name);
         }
 
         [Fact]
         public void CanSetNameFromReference()
         {
-            var book1 = GetBook("Book 1");
+            var book1 = GetBook("InMemoryBook 1");
             SetName(book1, "New Name");
 
             Assert.Equal("New Name", book1.Name);
@@ -89,30 +89,30 @@ namespace XUnitCSharpFundamentals
         [Fact]
         public void GetBookReturnsDifferentObjects()
         {
-            var book1 = GetBook("Book 1");
-            var book2 = GetBook("Book 2");
+            var book1 = GetBook("InMemoryBook 1");
+            var book2 = GetBook("InMemoryBook 2");
 
-            Assert.Equal("Book 1", book1.Name);
-            Assert.Equal("Book 2", book2.Name);
+            Assert.Equal("InMemoryBook 1", book1.Name);
+            Assert.Equal("InMemoryBook 2", book2.Name);
             Assert.NotSame(book1, book2);
         }
 
         [Fact]
         public void TwoVarsCanReferencesSameObject()
         {
-            var book1 = GetBook("Book 1");
+            var book1 = GetBook("InMemoryBook 1");
             var book2 = book1;
 
             Assert.Same(book1, book2);
         }
 
 
-        Book GetBook(string name)
+        InMemoryBook GetBook(string name)
         {
-            return new Book(name);
+            return new InMemoryBook(name);
         }
 
-        void SetName(Book book, string name)
+        void SetName(InMemoryBook book, string name)
         {
             book.Name = name;
         }
